% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_two_numbers.R
\name{subtract_two_numbers}
\alias{subtract_two_numbers}
\title{Subtract two numbers}
\usage{
subtract_two_numbers(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The subtraction of \code{y} from \code{x}.
}
\description{
Subtract two numbers
}
\examples{
subtract_two_numbers(1, 1)
subtract_two_numbers(10, 1)

}
